Move updateNetworkIface from libvirt-dbus.js to libvirt-xml-update.js
