+++ src/libvirt-xml-update.js
@@ -1,5 +1,5 @@
 import { getElem, getSingleOptionalElem } from './libvirt-common.js';
-import { getNextAvailableTarget } from './helpers.js';
+import { getNextAvailableTarget, logDebug } from './helpers.js';
 
 export function updateDisk({ domXml, diskTarget, readonly, shareable, busType, existingTargets, cache }) {
     const domainElem = getElem(domXml);
@@ -246,6 +246,88 @@ export function updateBootOrder(domXml, devices) {
     tmp.appendChild(domainElem);
 
     return tmp.innerHTML;
+}
+
+/**
+ * Returns updated XML description of the network interface specified by mac address.
+ * @param  {String} domXml      Domain XML description.
+ * @param  {String} networkMac  MAC Address of the network interface we will update.
+ * @param  {String} state       Desired state; one of up/down.
+ * @return {String}             Updated XML description of the device we will update or null on error.
+ */
+export function updateNetworkIface({ domXml, networkMac, networkState, networkModelType, networkType, networkSource }) {
+    const parser = new DOMParser();
+    const xmlDoc = parser.parseFromString(domXml, "application/xml");
+
+    if (!xmlDoc) {
+        console.warn(`Can't parse dumpxml, input: "${domXml}"`);
+        return null;
+    }
+
+    const domainElem = xmlDoc.getElementsByTagName("domain")[0];
+    const devicesElem = domainElem.getElementsByTagName("devices")[0];
+    const interfaceElems = devicesElem.getElementsByTagName('interface');
+
+    if (interfaceElems) {
+        for (let i = 0; i < interfaceElems.length; i++) {
+            const interfaceElem = interfaceElems[i];
+            const macElem = getSingleOptionalElem(interfaceElem, 'mac');
+            if (macElem === undefined)
+                return null;
+            const mac = macElem.getAttribute('address');
+
+            if (mac !== networkMac)
+                continue;
+
+            if (networkState) {
+                let linkElem = getSingleOptionalElem(interfaceElem, 'link');
+                if (linkElem === undefined) {
+                    const doc = document.implementation.createDocument('', '', null);
+                    linkElem = doc.createElement('link');
+                    interfaceElem.appendChild(linkElem);
+                }
+                linkElem.setAttribute('state', networkState);
+            }
+
+            const typeChanged = networkType !== interfaceElem.getAttribute('type', networkType);
+            if (networkType) {
+                interfaceElem.setAttribute('type', networkType);
+            }
+
+            if (networkSource && networkType) {
+                let sourceElem = getSingleOptionalElem(interfaceElem, 'source');
+                // Source elements of different iface types might contain differently named attributes,
+                // so we delete whole element and create a new one
+                if (typeChanged && sourceElem) {
+                    sourceElem.remove();
+                    sourceElem = undefined;
+                }
+                if (!sourceElem) {
+                    sourceElem = document.createElement("source");
+                    interfaceElem.appendChild(sourceElem);
+                }
+                if (networkType === 'network')
+                    sourceElem.setAttribute('network', networkSource);
+                else if (networkType === 'direct')
+                    sourceElem.setAttribute('dev', networkSource);
+                else if (networkType === 'bridge')
+                    sourceElem.setAttribute('bridge', networkSource);
+            }
+
+            if (networkModelType) {
+                const modelElem = getSingleOptionalElem(interfaceElem, 'model');
+                modelElem.setAttribute('type', networkModelType);
+            }
+
+            const returnXML = (new XMLSerializer()).serializeToString(interfaceElem);
+
+            logDebug(`updateNetworkIface: Updated XML: "${returnXML}"`);
+
+            return returnXML;
+        }
+    }
+    console.warn("Can't update network interface element in domXml");
+    return null;
 }
 
 /*
